Updated, commented, and seperated two models (SSIM and ORB)
